[refactor]: Rename model to expert
