+++ src/main/java/ladder/domain/expert/Direction.java
@@ -1,4 +1,4 @@
-package ladder.domain.model;
+package ladder.domain.expert;
 
 public enum Direction {
     LEFT(-1),
