+++ src/main/java/ladder/domain/expert/Line.java
@@ -1,4 +1,4 @@
-package ladder.domain.model;
+package ladder.domain.expert;
 
 import ladder.dto.LineDto;
 import ladder.exception.BadPositionException;
