+++ src/main/java/ladder/domain/expert/Person.java
@@ -1,4 +1,4 @@
-package ladder.domain.model;
+package ladder.domain.expert;
 
 import ladder.dto.PersonDto;
 import ladder.exception.BadNameException;
