+++ src/main/java/ladder/domain/expert/Point.java
@@ -1,4 +1,4 @@
-package ladder.domain.model;
+package ladder.domain.expert;
 
 import ladder.dto.PointDto;
 import ladder.exception.ConsecutiveDirectionException;
