+++ src/main/java/ladder/domain/expert/ValidatedResults.java
@@ -1,4 +1,4 @@
-package ladder.domain.model;
+package ladder.domain.expert;
 
 import ladder.exception.BadResultsException;
