+++ src/test/java/ladder/domain/expert/LineTest.java
@@ -1,4 +1,4 @@
-package ladder.domain.model;
+package ladder.domain.expert;
 
 import ladder.dto.PointDto;
 import ladder.exception.BadPositionException;
