+++ src/test/java/ladder/domain/expert/PersonTest.java
@@ -1,4 +1,4 @@
-package ladder.domain.model;
+package ladder.domain.expert;
 
 import ladder.dto.PersonDto;
 import ladder.exception.BadNameException;
