+++ src/test/java/ladder/domain/expert/ValidatedResultsTest.java
@@ -1,4 +1,4 @@
-package ladder.domain.model;
+package ladder.domain.expert;
 
 import ladder.exception.BadResultsException;
 import org.junit.jupiter.api.DisplayName;
